refactor(vital_records): move package task to subpackage
